feat: add lastProgress to VideoProgressSlider for selection cancellation

+++ app/src/view/video_view.cpp
@@ -628,13 +628,13 @@ void VideoView::registerMpvEvent() {
         case MpvEventEnum::UPDATE_DURATION:
             if (this->seekingRange == 0) {
                 this->rightStatusLabel->setText(misc::sec2Time(mpv.duration));
-                this->osdSlider->setProgress(mpv.playback_time / mpv.duration);
+                this->osdSlider->setProgress((float)mpv.playback_time / mpv.duration);
             }
             break;
         case MpvEventEnum::UPDATE_PROGRESS:
             if (this->seekingRange == 0) {
                 this->leftStatusLabel->setText(misc::sec2Time(mpv.video_progress));
-                this->osdSlider->setProgress(mpv.playback_time / mpv.duration);
+                this->osdSlider->setProgress((float)mpv.playback_time / mpv.duration);
             }
             break;
         case MpvEventEnum::END_OF_FILE:
